debian: Allow use of secondary groups
When '-g' is passed, freeswitch drops all other groups except for the
given group.  This impacts people who depend on FS having access to
resources that would be allowed by membership to those other groups.

It was possible to override this by setting DAEMON_ARGS in
/etc/default/freeswitch, but we'll go ahead and make this the default.

Since freeswitch uses the primary group of a user when `-g` is
omitted, we'll just omit it, and do similarly when setting the
ownership of our directory in /var/run.

Edited-by: Travis Cross <tc@traviscross.com>