Support optional debian parallel builds
With this commit we allow getting `make -jN` behavior by setting
`parallel=N` in `DEB_BUILD_OPTIONS`.  Setting `parallel=` will result
in `make -j` being run (i.e. no limit on the number of jobs).  The
default is to run one job at a time; this is still the safest bet when
building all modules.

Thanks to Peter Wu for suggesting a different way to do this.

ref: https://www.debian.org/doc/debian-policy/ch-source.html#s-debianrules-options